+++ changelog.md
@@ -5,6 +5,12 @@
 <br>
 <br>
 
+## Sep 17 2018
+- New Emotes panel now available.<br>
+  Press `F5` to open/close or `/emotes`.
+
+&nbsp;
+
 ## Sep 13 2018
 - New Emotes!
 - Use `/emote` to list all the options.Can be used by anyone!
@@ -12,6 +18,7 @@
 - Extensions: new `spectate` event.
 </ul>        
 
+&nbsp;
 
 ## Ago 23 2018
 - Default Theme is now Realistic Sprites.